+++ ospf6d/ospf6_interface.c
@@ -30,6 +30,7 @@
 #include "ospf6d.h"
 #include "ospf6_bfd.h"
 #include "ospf6_zebra.h"
+#include "ospf6_tlv.h"
 #include "ospf6_gr.h"
 #include "lib/json.h"
 #include "ospf6_proto.h"
